fix confused utxorescan
